% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{life}
\alias{life}
\title{Run Conway's Game of Life.}
\usage{
life(size, ngen = 1000, update.freq = 10)
}
\arguments{
\item{size}{(numeric) the size of the board (the length of its sides).}

\item{ngen}{(numeric) the number of generations to run.}

\item{update.freq}{(numeric) specifies the number of generations
between plot updates (since plot updates are by far the slowest here,
plotting results too often can become quite slow).}
}
\description{
\code{life} runs Conway's Game of Life, using standard rules.
}

